[ 이태훈 ] 리팩토링 : 게시판 이름,설명
