+++ src/views/CalendarPage/CalendarPage.js
@@ -7,6 +7,7 @@ import Header from '../Header/Header';
 import Quick from '../Quick/Quick';
 function CalendarPage(props) {
   const dispatch = useDispatch();
+
   useEffect(() => {
     dispatch(getScholar())
       .then((response) => { })
