+++ src/views/PostPage/Post.js
@@ -9,6 +9,7 @@ const PostView = loadable(() => import('../../components/post/PostView'));
 const PostList = loadable(() => import('../../components/post/PostList'));
 const PostUpdate = loadable(() => import('../../components/post/PostUpdate'));
 function Post({ match }) {
+
   return (
     <>
       <Header />
